+++ feature/main/src/main/java/com/note/feature/main/MainScreen.kt
@@ -19,7 +19,7 @@ import org.koin.androidx.compose.koinViewModel
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
 fun MainScreen(
-    onNavigateToDetail: (Int?) -> Unit,
+    onNavigateToDetail: (Long?) -> Unit,
     viewModel: MainViewModel = koinViewModel()
 ) {
     val state by viewModel.uiState.collectAsState()
